+++ extensions/note-outline/__tests__/Outline.test.tsx
@@ -1,1 +1,0 @@
-test.todo('renders outline correctly');